### 2025-04-02 #7992 @editorialbot
[PRE REVIEW]: PANKH: An unsteady potential flow solver for hovering airfoils
**Submitting author:** @coding4Acause (<a href="http://orcid.org/0009-0003-6701-5477">Rohit  Chowdhury</a>)
**Repository:** https://github.com/coding4Acause/PANKH
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
