+++ JOSS/paper.md
@@ -16,67 +16,43 @@ authors:
   - name: Ashish Pathak 
     orcid: 0000-0002-6608-1830
     affiliation: 1
-  - name: Department of Mechanical Engineering, Indian Institute of Technology, Jodhpur,   Rajasthan, India
+affiliations:
   - index: 1
-date: 28 March 2025
+    name: Department of Mechanical Engineering, Indian Institute of Technology, Jodhpur, Rajasthan, India
+date: 4 April 2025
 bibliography: paper.bib
 ---
 
 # Summary
-Aerodynamic analysis of unsteady airfoils plays a crucial role in understanding the flight mechanics of birds, insects, and micro air vehicles (MAVs). While computational fluid dynamics (CFD) provides detailed insights, lower-fidelity models such as potential flow solvers remain valuable due to their efficiency and ability to capture key aerodynamic phenomena. In particular, two-dimensional (2D) inviscid simulations serve as an essential tool for rapid analysis of unsteady aerodynamics, reducing computational costs while retaining fundamental flow physics.
+Analyzing the aerodynamics of unsteady airfoils is vital for decoding the flight dynamics of birds, insects, fixed-wing aircrafts and micro air vehicles (MAVs). Although computational fluid dynamics (CFD) delivers in-depth results, lower-fidelity approaches like potential flow solvers offer significant value due to their speed and capacity to model essential aerodynamic effects.
 
-This work presents **PANKH (Panel Analysis of uNsteady Kinematics of Hovering airfoils)**, an open-source C++ solver implementing the unsteady vortex panel method for analyzing the aerodynamic forces on airfoils undergoing arbitrary motion. The solver discretizes the airfoil surface into panels with a piecewise linear vortex distribution and enforces the no-penetration boundary condition at control points. The Kutta condition ensures smooth trailing-edge flow, while the Kelvin circulation theorem governs the shedding of wake vortices. The wake is dynamically modeled using constant-strength vortex panels, which are convected as discrete point vortices in subsequent time steps. The resulting system of equations is solved using the Eigen library, providing an efficient and robust numerical framework.
+We introduce PANKH **(Panel Analysis of uNsteady Kinematics of Hovering airfoils)**, an open-source C++ tool that employs the unsteady vortex panel method to evaluate aerodynamic forces on airfoils in arbitrary motion. Its flexible, modular design enables users to specify custom kinematic patterns, ideal for exploring bio-inspired flapping flight and gust responses.
 
-PANKH computes unsteady aerodynamic forces such as lift, drag, and moment coefficients for airfoils undergoing pitching, plunging, or combined motions. The modular structure of the code allows users to define custom kinematic profiles, making it suitable for studying bio-inspired flapping flight, rotor aerodynamics, and gust interactions. The solver is designed for high computational efficiency, making it a valuable tool for researchers and engineers investigating unsteady aerodynamics without the computational expense of full CFD simulations.
-
-The source code, validation cases, and documentation are available on GitHub, ensuring accessibility and reproducibility. Future developments may include expanded input handling, additional wake modeling techniques, and extensions to three-dimensional flow analysis.
-
+The solver’s source code, validation examples, and documentation are hosted on GitHub, guaranteeing ease of access and replicability. Upcoming enhancements might feature broader input capabilities, refined wake modeling approaches, and the integration of alternative singularity distributions.
 
 # Statement of Need
 
[…]
