+++ JOSS/paper.bib
@@ -1,8 +1,11 @@
 @book{dimitri,
 title={Unsteady Aerodynamics: Potential and Vortex Methods},
-author={Dimitriadis, Grigorios},
+author={Grigorios Dimitriadis},
 year={2023},
-publisher={John Wiley \& Sons}
+publisher={John Wiley & Sons, 2023},
+ISBN ={978-1-119-76247-8},
+Length={576 pages},
+url={https://books.google.co.in/books?id=VnbiEAAAQBAJ}
 }
 
 @article{basu_1978,
@@ -12,7 +15,8 @@ publisher={John Wiley \& Sons}
   year={1978},
   volume={87},
   pages={159 - 178},
-  url={https://api.semanticscholar.org/CorpusID:123609694}
+  doi={10.1017/S0022112078002980},
+  url={https://www.cambridge.org/core/product/identifier/S0022112078002980/type/journal_article}
 }
 
 @inproceedings{prosser2011applicability,
@@ -20,7 +24,9 @@ publisher={John Wiley \& Sons}
   author={Prosser, Daniel and Crassidis, Agamemnon and Ghosh, Amitabha and Roemer, Mike},
   booktitle={AIAA Atmospheric Flight Mechanics Conference},
   pages={6392},
-  year={2011}
+  year={2011},
+  doi={10.2514/6.2011-6392},
+  url={https://arc.aiaa.org/doi/10.2514/6.2011-6392}
 }
 @article{vezza1985method,
   title={A method for predicting unsteady potential flow about an aerofoil},
@@ -30,25 +36,20 @@ publisher={John Wiley \& Sons}
   number={4},
   pages={347--356},
   year={1985},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi= {10.1002/fld.1650050405},
+  url={https://onlinelibrary.wiley.com/doi/10.1002/fld.1650050405}
 }
 @book{katz2001low,
   title={Low-speed aerodynamics},
   author={Katz, Joseph and Plotkin, Allen},
   volume={13},
   year={2001},
-  publisher={Cambridge university press}
+  publisher={Cambridge university press},
+  doi={10.1115/1.1669432},
+  url={https://asmedigitalcollection.asme.org/fluidsengineering/article/126/2/293/458666/LowSpeed-Aerodynamics-Second-Edition}
 }
 
-@article{hess1967calculation,
-  title={Calculation of potential flow about arbitrary bodies},
-  author={Hess, John L and Smith, A\_M O},
-  journal={Progress in Aerospace Sciences},
-  volume={8},
-  pages={1--138},
-  year={1967},
-  publisher={Pergamon}
-}
 @article{giesing1968nonlinear,
   title={Nonlinear two-dimensional unsteady potential flow with lift.},
   author={Giesing, Joseph P},
@@ -56,46 +57,17 @@ publisher={John Wiley \& Sons}
   volume={5},
   number={2},
   pages={135--143},
-  year={1968}
-}
-@article{simpson2013induced,
-  title={Induced-drag calculations in the unsteady vortex lattice method},
-  author={Simpson, Robert JS and Palacios, Rafael and Murua, Joseba},
-  journal={AIAA journal},
-  volume={51},
-  number={7},
-  pages={1775--1779},
-  year={2013},
-  publisher={American Institute of Aeronautics and Astronautics}
-}
-
-@article{roccia2013modified,
-  title={Modified unsteady vortex-lattice method to study flapping wings in hover flight},
-  author={Roccia, Bruno A and Preidikman, Sergio and Massa, Julio C and Mook, Dean T},
-  journal={AIAA journal},
-  volume={51},
[…]
