### 2025-04-04 #7992 @coding4Acause
Thank you, @arfon . I’ve just fixed the DOIs that weren’t functioning correctly. Regarding the editorial review flag, I’ll wait until next week for a response.
