### 2025-06-30 #7992 @coding4Acause
@kyleniemeyer,

Thank you for your valuable feedback on the **PANKH** (Panel Analysis of uNsteady Kinematics of Hovering airfoils) software submission. We appreciate your concern regarding its utility for research. Below, we address this by highlighting **PANKH**’s unique features, its relevance to current research needs, and its potential to advance studies in unsteady aerodynamics, particularly for low-speed flows.

PANKH is a specialized, open-source C++ software designed to model unsteady potential flow for hovering airfoils, addressing a critical gap in aerodynamic analysis tools, as outlined below:

1. **Rapid Analysis for Preliminary Design**
   In designing fixed-wing aircraft, modern flexible aircraft, micro-air vehicles (MAVs), and unmanned aerial vehicles (UAVs), the aerodynamic performance of airfoil cross-sections is crucial. Traditional methods for unsteady or quasi-steady aerodynamic loads, such as wind tunnel tests or high-fidelity computational fluid dynamics (CFD) simulations (e.g., Navier-Stokes or Lattice-Boltzmann methods), are time-consuming and computationally intensive. **PANKH** offers a low-fidelity, computationally efficient alternative, delivering results in seconds on standard hardware. This enables researchers to:
   - Perform rapid parametric studies during preliminary design.
   - Evaluate lift and drag polars for unsteady airfoil motions (e.g., pitching or plunging).
   - Optimize airfoil shapes and kinematics for applications like fixed-wing aircraft and flapping-wing MAVs.
   By providing fast, first-order solutions, **PANKH** reduces the need for costly experimental setups or high-fidelity CFD in early-stage design, supporting rapid parametric studies, optimization, and real-time simulation of unsteady flows.

2. **Specialization for Unsteady Hovering Aerodynamics**
   To underscore **PANKH**’s unique contribution, we highlight the limitations of existing open-source tools, which emphasize the need for a specialized tool for **unsteady aerodynamics** of hovering airfoils. Numerous low- to medium-fidelity panel method solvers, such as MIT’s [**XFOIL**](http://web.mit.edu/drela/Public/web/xfoil/) by Mark Drela, NASA’s [**FoilSim III**](https://www.grc.nasa.gov/WWW/k-12/airplane/FoilElemJS/FoilSim.html) and [**JavaFoil**](https://www.mh-aerotools.de/airfoils/javafoil.htm), are tailored for steady flow analysis and are limited in their ability to capture unsteady flow phenomena effectively. In contrast, PANKH is purpose-built for unsteady aerodynamics of hovering airfoils, offering advanced features for applications like flapping-wing MAVs, ornithopters, and hovering rotorcraft. Key features include:
   - *Time-Stepping Wake Model*: **PANKH** incorporates a sophisticated time-stepping wake model to capture the complex dynamics of vortex shedding, a critical aspect of unsteady aerodynamics. Shed vortices convect, interact, and induce velocities on neighboring vortices and the airfoil, profoundly influencing the pressure field and aerodynamic loads. Some tools use simplified models, such as representing the latest shed vortex as a point vortex, as seen in the solver [**ULVPC**](https://doi.org/10.2514/6.2011-6392). This approach compromises physical realism. In contrast, **PANKH** adopts the pioneering method of [Basu and Hancock, 1978](https://doi.org/10.1017/S0022112078002980). It models the current wake vortex as a constant-strength vortex panel. This method ensures a more accurate representation of wake behavior. **PANKH**’s efficient time-stepping framework adeptly handles the numerical challenges of stability, vorticity conservation, and computational complexity, delivering precise simulations of unsteady wake effects for hovering and flapping motions.

   - *Unsteady Kutta Condition*: **PANKH** implements an advanced Kutta condition tailored for unsteady flows, ensuring realistic flow behavior at the airfoil’s trailing edge. The Kutta condition enforces smooth flow departure, preventing unphysical pressure singularities and setting accurate circulation for lift prediction. In unsteady flows, rapid changes in angle of attack and vortex shedding create time-varying circulation, requiring a dynamic Kutta condition to maintain vorticity continuity and zero pressure difference at the trailing edge. PANKH’s formulation, inspired by Wagner’s analytical solution, achieves this, enabling precise lift and drag predictions for flapping-wing MAVs and hovering rotorcraft, distinguishing it from other open-source tools.
   - *Real-Time Visualization*: The software provides real-time visualization of wake dynamics, enabling researchers to study unsteady flow phenomena like vortex-airfoil interactions or lift transients.

3. **Demonstrated Research Impact**
    PANKH has been successfully utilized in a [study](https://www.researchgate.net/publication/393158131_Development_of_an_unsteady_vortex_panel_method_for_a_flapping_airfoil) submitted to the Fluid Mechanics and Fluid Power (FMFP) conference, held December 21–23, 2024, in Aligarh Muslim University (AMU), India, to investigate unsteady aerodynamic loads on hovering airfoils. The findings offered significant insights into lift and drag variations under dynamic conditions, underscoring PANKH's practical value. As an open-source tool, planned for release via platforms like the Journal of Open Source Software (JOSS), PANKH will reach a broader research community, encouraging adoption and collaboration. We anticipate increased citations as researchers leverage its capabilities for bio-inspired flight, rotorcraft, and related studies.
