### 2025-08-22 #7992 @piyueh
@daavid00 @anand-me @stebiondi @andrewgiuliani @jamesafful @ma-sadeghi @jakobhaervig @knstmrd Thank you for signing up as JOSS reviewers! Since your reviewer profiles indicate expertise in computational fluid dynamics and C/C++, would you be willing to review this submission for JOSS?

The submission focuses on potential flow/inviscid flow/panel methods and is implemented in C/C++. If you haven’t reviewed for JOSS before, we have detailed guidelines and a checklist available at [https://joss.readthedocs.io/](https://joss.readthedocs.io/) to help you get started.

We are aiming to recruit at least two reviewers, ideally three.
