### 2025-08-22 #7992 @ma-sadeghi
Thanks for the invite. I'd be happy to, but I'm busy until end of October.
