### 2025-08-26 coding4Acause/PANKH#1 @daavid00
Missing Contributing.md
Regarding the review point:

- **Community guidelines:** Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

this is missing in the repository (see https://contributing.md/example/ for an example).

Related to openjournals/joss-reviews#8865
