### 2025-08-26 coding4Acause/PANKH#2 @daavid00
Missing automated tests
Regarding the review point:

- **Automated tests**: Are there automated tests or manual steps described so that the functionality of the software can be verified?

this seems to be missing. One way to handle this is to have a continuous integration (.github/workflows/CI.yml) file that runs via Github action to test the installation of prerequisites, compilation of the software, and run the tests (see this nice tutorial for an example https://www.incredibuild.com/blog/using-github-actions-with-your-c-project).

Related to https://github.com/openjournals/joss-reviews/issues/8865
