### 2025-08-26 916532f
Fix to error and warning in compiling with clang/macOS
