### 2025-08-26 coding4Acause/PANKH#4 @daavid00
macOS support
The PR https://github.com/coding4Acause/PANKH/pull/3 allows to compile in macOS.

I would recommend to add in the documentation that eigen and gnuplot can be installed via brew in macOS (e.g., `brew install eigen`), and for compilation one needs to add the path to eigen (i.e., `g++ -I/opt/homebrew/Cellar/eigen/3.4.0_1/include/eigen3/ -Iinclude src/*.cpp -o PANKH_solver`) and for execution this needs to be without empty space (`./PANKH_solver` instead of `./ PANKH_solver`).

In addition, I would recommend to use `set terminal qt` instead to `set terminal x11` [here](https://github.com/coding4Acause/PANKH/blob/main/src/gnuplot.cpp#L5) and [over here](https://github.com/coding4Acause/PANKH/blob/main/src/gnuplot.cpp#L38), which allows to visualise the results in macOS (this since brew install gnuplot comes with qt).

Related to https://github.com/openjournals/joss-reviews/issues/8865
