### 2025-08-26 f67a889
[DO NOT MERGE] Paper review comments
