+++ JOSS/paper.md
@@ -1,7 +1,7 @@
 ---
 title: 'PANKH: An unsteady potential flow solver for hovering airfoils'
 tags:
-  - Unsteady aerodynamics
+  - Unsteady Aerodynamics
   - Potential Flows
   - Vortex Panel
   - Kutta Condition
@@ -24,7 +24,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-Analyzing the aerodynamics of unsteady airfoils is vital for decoding the flight dynamics of birds, insects, fixed-wing aircrafts and micro air vehicles (MAVs). Although computational fluid dynamics (CFD) delivers in-depth results, lower-fidelity approaches like potential flow solvers offer significant value due to their speed and capacity to model essential aerodynamic effects.
+Analyzing the aerodynamics of unsteady airfoils is vital for decoding the flight dynamics of birds, insects, fixed-wing aircrafts and micro air vehicles (MAVs). Although computational fluid dynamics (CFD) delivers in-depth results, lower-fidelity approaches like potential flow solvers offer significant value due to their speed and capacity to model essential aerodynamic effects (TODO: essential aerodynamic effects such as? Here you could mention a few of them).
 
 We introduce PANKH **(Panel Analysis of uNsteady Kinematics of Hovering airfoils)**, an open-source C++ tool that employs the unsteady vortex panel method to evaluate aerodynamic forces on airfoils in arbitrary motion. Its flexible, modular design enables users to specify custom kinematic patterns, ideal for exploring bio-inspired flapping flight and gust responses.
 
@@ -32,7 +32,7 @@ The solver’s source code, validation examples, and documentation are hosted on
 
 # Statement of Need
 
-For low-speed aerodynamic applications, where viscous effects are primarily confined to thin boundary layers and wakes, potential flow solvers provide a computationally efficient alternative to high-fidelity CFD. These solvers approximate the flow as inviscid, incompressible, and irrotational, reducing complexity while still offering accurate lift estimations. By solving Laplace’s equation with appropriate boundary conditions, potential flow methods enable rapid aerodynamic analysis, making them particularly useful for preliminary design, scaling studies, and parametric investigations. Despite their advantages, open-source tools for unsteady potential flow analysis remain scarce. While [XFOIL](https://web.mit.edu/drela/Public/web/xfoil/), developed by Mark Drela, is a widely used tool for steady-state airfoil aerodynamics, it lacks the capability to handle unsteady effects. Existing research codes for unsteady vortex panel methods are often inaccessible, undocumented, or not actively maintained.
[…]
