### 2025-08-26 coding4Acause/PANKH#7 @daavid00
Parsing the input values instead of recompiling every time
To avoid the recompilation of the executable every time the parameters are changed in [constants.cpp](https://github.com/coding4Acause/PANKH/blob/main/src/constants.cpp), then one can parse these values from an input file (e.g., see [this nice tutorial](https://dev.to/uponthesky/c-making-a-simple-json-parser-from-scratch-250g)  for parsing from JSON files).

While nice for users, this is not a requirement for https://github.com/openjournals/joss-reviews/issues/8865
