### 2025-08-28 ce44d7d
Update impulsive_start.cpp
Signed-off-by: Rohit Chowdhury <d22me102@iitj.ac.in>
