### 2025-08-31 374dc99
update .h and .cpp files with current working version
