+++ src/geometry.cpp
@@ -1,12 +1,8 @@
 #include "geometry.h"
-#include "kinematics.h"
-#include "constants.h"
-#include "VectorOperations.h"
 #include <fstream>
 
-
 /* this function returns the upper and lower coordinates (dimensionalised) of a NACA 4 digit series airfoil for any x/c. */
-MatrixXd geometry(double x_c)
+MatrixXd geometry(double x_c, double c, double q, double p, int trailing_edge_type, double t_m)
 {
 
     MatrixXd airfoil_points(3, 2);
@@ -23,13 +19,13 @@ MatrixXd geometry(double x_c)
         yc = (p * c * (2.0 * (1 - x_c) / (1 - q) - pow(((1 - x_c) / (1 - q)), 2.0)));
         der_yc = c * ((2.0 * p / (1.0 - q)) * (((1.0 - x_c) / (1.0 - q)) - 1.0));
     }
-    if (trailing_edge_type == 1.0) // 1 indicates open trailing edge
+    if (trailing_edge_type == 1) // 1 indicates open trailing edge
     {
-        t = c * (tm * (2.969 * sqrt(x_c) - 1.260 * (x_c)-3.516 * pow((x_c), 2) + 2.843 * pow((x_c), 3.0) - 1.015 * (pow((x_c), 4.0))));
+        t = c * (t_m * (2.969 * sqrt(x_c) - 1.260 * (x_c)-3.516 * pow((x_c), 2) + 2.843 * pow((x_c), 3.0) - 1.015 * (pow((x_c), 4.0))));
     }
     else
     {
-        t = c * (tm * (2.980 * sqrt(x_c) - 1.320 * (x_c)-3.286 * pow((x_c), 2) + 2.441 * pow((x_c), 3.0) - 0.815 * (pow((x_c), 4.0))));
+        t = c * (t_m * (2.980 * sqrt(x_c) - 1.320 * (x_c)-3.286 * pow((x_c), 2) + 2.441 * pow((x_c), 3.0) - 0.815 * (pow((x_c), 4.0))));
     }
     if (p == 0.0) //[SYMMETRIC AIRFOIL]
     {
@@ -52,8 +48,7 @@ MatrixXd geometry(double x_c)
     return airfoil_points;
 }
 
-
-void nodal_coordinates_initial(VectorXd &x0, VectorXd &y0) // this function basically discretizes the chord in a nonlinear way [COSINE CLUSTERING]
+void nodal_coordinates_initial(int n,double c, double q, double p, int trailing_edge_type, double t_m, VectorXd &x0, VectorXd &y0) // this function basically discretizes the chord in a nonlinear way [COSINE CLUSTERING]
 {
     double theta;
     MatrixXd panel_points(3, 2);
@@ -70,7 +65,7 @@ void nodal_coordinates_initial(VectorXd &x0, VectorXd &y0) // this function basi
         {                                    // storing lower coordinates first
             theta = (i - 0.5) * delta_theta; // cosine clustering
             x_nd = 0.5 * (1 - cos(theta));
-            panel_points = geometry(x_nd);
+            panel_points = geometry(x_nd, c, q, p, trailing_edge_type, t_m);
             x0(n / 2 - i) = panel_points(1, 0); // x_lower
             y0(n / 2 - i) = panel_points(1, 1); // y_lower
         }
@@ -78,7 +73,7 @@ void nodal_coordinates_initial(VectorXd &x0, VectorXd &y0) // this function basi
         {                                    // storing upper coordinates
             theta = (i - 0.5) * delta_theta; // cosine clustering
             x_nd = 0.5 * (1 - cos(theta));
-            panel_points = geometry(x_nd);
[…]
