### 2025-08-31 02a82ea
Update impulsive_start.cpp
Signed-off-by: Rohit Chowdhury <d22me102@iitj.ac.in>
