### 2025-08-31 bd56688
update with execution time calculation
