### 2025-08-31 36d51a7
update after laoding input json file
