### 2025-09-01 2b4bf4a
test critical functions
