### 2025-09-01 72db4d5
Update CI.yml
Signed-off-by: Rohit Chowdhury <d22me102@iitj.ac.in>
