### 2025-09-01 83f49a1
Update CI.yml
Signed-off-by: Rohit Chowdhury <d22me102@iitj.ac.in>
