+++ contributing.md
@@ -1,0 +1,185 @@
+# Contributing to PANKH
+
+First off, thank you for considering contributing to **PANKH**!   
+Your contributions—whether in the form of code, documentation, bug reports, feature requests, or testing—are all highly appreciated.
+
+>  If you're not ready to contribute code but want to support the project:
+> -  Star the repository  
+> -  Share it with others in your field  
+> -  Cite it in your academic work  
+> -  Use it in your classroom or lab  
+
+---
+
+## Table of Contents
+
+- [Code of Conduct](#code-of-conduct)  
+- [How to Contribute](#how-to-contribute)  
+  - [Reporting Bugs](#reporting-bugs)  
+  - [Suggesting Features](#suggesting-features)  
+  - [Adding New Code or Examples](#adding-new-code-or-examples)  
+  - [Improving Documentation](#improving-documentation)  
+- [Style Guide](#style-guide)  
+- [Commit Messages](#commit-messages)  
+- [Setting Up Locally](#setting-up-locally)  
+- [Running Tests](#running-tests)  
+- [Project Vision and Contribution Opportunities](#project-vision-and-contribution-opportunities)
+
+
+
+---
+
+## Code of Conduct
+
+Please read our [Code of Conduct](CODE_OF_CONDUCT.md) before contributing. By participating in this project, you agree to follow its terms.
+
+---
+
+## How to Contribute
+
+### Reporting Bugs
+
+If you encounter a bug:
+
+- Check if it’s already reported in [Issues](https://github.com/coding4Acause/PANKH/issues)
+- If not, [open a new issue](https://github.com/coding4Acause/PANKH/issues/new)
+- Include:
+  - What you expected to happen
+  - What actually happened
+  - Steps to reproduce it
+  - System info: OS, compiler version, etc.
+  - Input files, if possible
+
+### Suggesting Features
+
+If you have an idea for a feature:
+
+- First, check [existing issues](https://github.com/coding4Acause/PANKH/issues)
+- If it's not there, open a new issue with a clear title and description
+- Explain why it would be useful, and any ideas on implementation
+
+
+### Adding New Code or Examples
+
+If you're submitting new code (e.g. motion modules, solvers, visualization):
+
+- Fork the repository
+- Create a new branch (`feature/your-feature-name`)
+- Keep functions modular and avoid hardcoding
+- If you're adding examples (e.g. sudden acceleration, pitch–plunge):
+  - Place input files in `examples/your_case/`
+  - Write a short README for the new example
+  - Make sure it runs with `./PANKH_solver input.json`
+
+### Improving Documentation
+
+You can improve:
+
+- The `README.md` (installation, usage, references)
+- The `examples/` documentation
+- The input JSON format explanation
+- Add doc comments in the code
+
+---
+
+## Style Guide
+
+- Use **C++11 or newer**
+- No classes/structs (for now) – keep procedural
+- Maintain consistent formatting:
+  - 4-space indentation
+  - `snake_case` for variable names
+  - Use `Eigen::VectorXd`, `Eigen::MatrixXd`, etc. when applicable
+- Keep the `main.cpp` clean – parse input in a separate file
+- Place `.h` files in `includes/`, `.cpp` in `src/`
+
+---
+
+## Commit Messages
+
+Please follow this format:
+
+```
+<type>: <short summary>
+
+<body - optional, wrap at 72 chars>
+```
+
+Example:
+
+```
+fix: correct vortex strength update in unsteady solver
+
+The Kutta condition wasn’t enforced properly for high reduced frequencies.
+```
+
+Types: `fix`, `feat`, `docs`, `refactor`, `test`, `chore`
[…]
