### 2025-09-02 a850b18
docs: update README.md with project badges and CI status
