### 2025-09-03 fb3d8cf
docs(joss): update paper.md with refined statement of need
