### 2025-09-03 7ff4c65
docs(joss): update paper.md with refined statement of need
