### 2025-09-03 151c8f6
docs(joss): update paper.md with refined statement of need
