+++ JOSS/paper.md
@@ -35,7 +35,7 @@ The solver’s source code, validation examples, and comprehensive Doxygen-gener
 
 For low-speed aerodynamic applications, where viscous effects are predominantly confined to thin boundary layers and wakes, potential flow solvers offer a computationally efficient alternative to high-fidelity computational fluid dynamics (CFD) and experimental approaches. High-fidelity CFD simulations, executed on high-performance computing (HPC) clusters comprising multiple nodes and cores, often demand days of runtime due to their intensive computational requirements and parallel processing across distributed architectures. Similarly, experimental methods require months to years for designing setups, procuring flow visualization equipment, and conducting tests. In contrast, PANKH, a modular C++ based open-source solver, delivers accurate aerodynamic load estimations within minutes on standard single-core systems, making it ideal for rapid prototyping and analysis. 
 
-Low-to-medium fidelity solvers like PANKH solve Laplace’s equation to model inviscid, incompressible, and irrotational flows, minimizing computational complexity while delivering precise lift estimations for aerodynamic analysis. By solving Laplace’s equation with appropriate boundary conditions, PANKH enable rapid aerodynamic analysis, making them particularly useful for preliminary design, scaling studies, and parametric investigations. Despite their advantages, open-source tools for unsteady potential flow analysis remain scarce. Established tools like [XFOIL](https://web.mit.edu/drela/Public/web/xfoil/), developed by Mark Drela, NASA's FoilSIM III, JavaFoil are tailored for steady-state aerodynamcis and lack the capability to address unsteady phenomena. Existing research codes for unsteady vortex panel methods are often proprietary, poorly documented, or no longer maintained. In contrast, PANKH is purpose-built for the unsteady aerodynamics of hovering airfoils, offering advanced features tailored for applications such as fixed-wing aircraft, flapping-wing micro-air vehicles (MAVs), ornithopters, and hovering rotorcraft. The software repository includes  validation cases comparing PANKH's results with experimental studies by [@anderson1998oscillating] and [@floryan2017scaling]. For three-dimensional wing design, understanding two-dimensional cross-sectional properties through airfoil analysis is critical for selecting optimal shapes. In real-world applications, PANKH holds significant potential for the aviation industry in its current form, with enhanced applicability when integrated with structural solvers to enable strongly coupled fluid-structure interaction analyses. Additionally, PANKH can serve as an effective tool for classroom visualizations, owing to its straightforward, modular, and procedural backend, which provides substantial flexibility for simulating diverse kinematic patterns.
[…]
