### 2025-09-03 4f06245
docs(joss): update paper.md with refined statement of need
