+++ JOSS/paper.bib
@@ -79,3 +79,25 @@ url={https://books.google.co.in/books?id=VnbiEAAAQBAJ}
   doi={10.1007/978-3-030-18319-6},
   url={http://link.springer.com/10.1007/978-3-030-18319-6}
 }
+@article{anderson1998oscillating,
+  title={Oscillating foils of high propulsive efficiency},
+  author={Anderson, Jamie M and Streitlien, K and Barrett, DS and Triantafyllou, Michael S},
+  journal={Journal of Fluid mechanics},
+  volume={360},
+  pages={41--72},
+  year={1998},
+  publisher={Cambridge University Press},
+  doi={10.1017/S0022112097008392},
+  url={https://doi.org/10.1017/S0022112097008392}
+}
+@article{floryan2017scaling,
+  title={Scaling the propulsive performance of heaving and pitching foils},
+  author={Floryan, Daniel and Van Buren, Tyler and Rowley, Clarence W and Smits, Alexander J},
+  journal={Journal of Fluid Mechanics},
+  volume={822},
+  pages={386--397},
+  year={2017},
+  publisher={Cambridge University Press},
+  doi={10.1017/jfm.2017.302},
+  url={https://doi.org/10.1017/jfm.2017.302}
+}
