### 2025-09-10 7f22774
docs: add platform-specific visualization setup instruction using gnuplot
