### 2025-09-10 d85b8bf
docs: add instructions in README.md for selecting Gnuplot terminals via input.json
