### 2025-09-10 e439a85
feat: enable user-selectable Gnuplot terminals
