### 2025-09-11 6187b2e
feat: add JSON configuration support for impulsive_start executable with user-defined Gnuplot terminal
