+++ examples/sudden_acceleration/impulsive_start.cpp
@@ -7,60 +7,34 @@
 #include <fstream>
 #include <cstdio>
 #include <chrono>
+#include "json.hpp"
 
+using json = nlohmann::json;
 using namespace Eigen;
 using namespace std;
 
 #define pi acos(-1)
-#define n 101 // the number of nodes [# panels=n-1]
+#define DEG2RAD pi / 180.0
 
-/* NACA 4 digit series airfoil parameters */
-const int ymc = 0;
-const int xmc = 0;
-const int tmax = 12;
-const int trailing_edge_type = 2; //[1-closed,else-open]
-
-double c = 0.2;     // chord length
-double Qinf = 10.0; // magnitude of the freestream velocity in m/sec
-double dt = (16.0 * c) / ((n-1) * Qinf);
-double tou_max = 50.0; // non-dimensional time
-double timemax = tou_max * c * 0.5 / Qinf;
-const int nsteps = timemax / dt; // max number of iterations or time steps.............
-
-int wakechoice = 0; //[0-free wake, 1-prescribed wake]
-
-/*In this problem we are rotating the airfoil by instantaneous angle of attack keeping the freestream direction parallel to X dirn.[INERTIAL FRAME]*/
-double Uinf = Qinf;
-double Vinf = 0.0;
-Vector2d freestream(Uinf, Vinf);
-
-// Motion parameters [needed for pitch/plunge cases]
-double k = 0.2;                                             // Reduced frequency
-double omega = (2.0 * k * Qinf) / c;                        // Angular Frequency of oscillation [rad/sec]
-double T = 2.0 * pi / omega;                                // Time Period of oscillation
-double h1 = 0.25 * c;                                       // Plunge amplitude
-double alpha1 = 15.0 * pi / 180.0 - atan2(2.0 * k * h1, c); // Pitch amplitude
-double h0 = 0.0;                                            // Mean plunge
-double alpha0 = 0.0 * pi / 180.0;                           // Mean pitch
-double phi_h = 0.0;                                         // Plunge phase
-double phi_alpha = (90.0 + phi_h) * pi / 180.0;             // Pitch phase
-
-// For newtonRaphson
-double tolerance = 1.e-5; // convergence criteria
-double epsilon = 1.e-8;   // increment
+// Following parameters are always zero since in IMPUSLISVE MOTION case it is assumed that there is no pitching and plunging.
 
 VectorXd normalize_2d(VectorXd v)
 {
-    double mag;
-    mag = sqrt(v(0) * v(0) + v(1) * v(1));
+    double mag = sqrt(v(0) * v(0) + v(1) * v(1));
+    if (mag < 1e-10)
+    { // Avoid division by zero
+        return VectorXd::Zero(2);
+    }
     v = v / mag;
     return v;
 }
-double magnitude(VectorXd a)
+double magnitude(const VectorXd &a)
 {
-    double mag;
-    mag = sqrt(a(0) * a(0) + a(1) * a(1));
-    return mag;
+    if (a.size() < 2)
+    {
+        throw invalid_argument("Vector must have at least 2 elements");
+    }
+    return sqrt(a(0) * a(0) + a(1) * a(1));
 }
 double dot(VectorXd a, VectorXd b)
 {
@@ -77,15 +51,13 @@ VectorXd cross(VectorXd a, VectorXd b)
     return cross_product;
 }
 
[…]
