### 2025-09-11 72e7d87
Update impulsive_start.cpp
Signed-off-by: Rohit Chowdhury <d22me102@iitj.ac.in>
