### 2025-09-11 8fd9b35
docs: add references and highlight PANKH's functionality
