+++ JOSS/paper.bib
@@ -101,3 +101,21 @@ url={https://books.google.co.in/books?id=VnbiEAAAQBAJ}
   doi={10.1017/jfm.2017.302},
   url={https://doi.org/10.1017/jfm.2017.302}
 }
+@inproceedings{drela1989xfoil,
+  title={XFOIL: An analysis and design system for low Reynolds number airfoils},
+  author={Drela, Mark},
+  booktitle={Low Reynolds number aerodynamics: proceedings of the conference notre dame, Indiana, USA, 5--7 June 1989},
+  pages={1--12},
+  year={1989},
+  organization={Springer},
+  doi={10.1007/978-3-642-84010-4_1},
+  url={https://link.springer.com/chapter/10.1007/978-3-642-84010-4_1}
+}
+
+@misc{eigenweb,
+  author       = {Gaël Guennebaud and Benoît Jacob and others},
+  title        = {Eigen v3: C++ Template Library for Linear Algebra},
+  year         = {2010},
+  howpublished = {\url{https://eigen.tuxfamily.org}},
+  note         = {Accessed September 2025}
+}
