### 2025-10-09 9f2c5f4
docs(joss): adresses reviewer commnet in issue #11
