+++ JOSS/paper.bib
@@ -119,3 +119,18 @@ url={https://books.google.co.in/books?id=VnbiEAAAQBAJ}
   howpublished = {\url{https://eigen.tuxfamily.org}},
   note         = {Accessed September 2025}
 }
+
+@misc{hepperle_javafoil,
+  title        = {JavaFoil - Airfoil Analysis Program},
+  author       = {Martin Hepperle},
+  year         = {2020},
+  note         = {\url{http://www.mh-aerotools.de/airfoils/javafoil.htm}}
+}
+
+@misc{nasa_foilsim,
+  title        = {FoilSim III – Interactive Airfoil Simulation},
+  author       = {NASA Glenn Research Center},
+  year         = {2022},
+  note         = {Available at \url{https://www.grc.nasa.gov/www/k-12/FoilSim3/}}
+}
+
