### 2025-10-09 a1a8088
docs(joss): adresses reviewer comments in issue #11
Signed-off-by: Rohit Chowdhury <d22me102@iitj.ac.in>
