### 2025-10-10 4bedf17
docs(JOSS):update paper.md and paper.pdf
