+++ JOSS/paper.md
@@ -111,7 +111,7 @@ Here, $\Gamma$ represents the total circulation in the flow domain enclosed by t
 
 This equation provides the $(n+1)^{\text{th}}$ constraint, required to solve for the n+1 unknowns in the system. 
 
-The resulting system, formulated as shown in **Figure 2**, is expressed as a single matrix equation and solved using appropriate routines from the *Dense Linear Problems and Decompositions* module of the Eigen library [[@eigenweb]](https://eigen.tuxfamily.org/index.php?title=Main_Page). In our implementation, we experimented with both the ColPivHouseholderQR and PartialPivLU classes. It was observed that both methods yielded comparable accuracy, while PartialPivLU was marginally faster than ColPivHouseholderQR in our specific use case.
+The resulting system, formulated as shown in **Figure 2**, is expressed as a single matrix equation and solved using the PartialPivLU decomposition solver available in the *Dense Linear Problems and Decompositions* module of the Eigen Library [[@eigenweb]](https://eigen.tuxfamily.org/index.php?title=Main_Page). 
 
 ![Equations formulated to determine the set of unknowns at a time instant $t_k$.](ax=b.jpg)
 
