### 2025-10-10 eaeeb33
add code of conduct adaapted from static version at contributor-covenant.org
