### 2025-10-10 6c4f644
update the run command
