+++ .github/workflows/CI.yml
@@ -31,5 +31,5 @@ jobs:
       - name: 🧪 Compile and run test
         run: |
           g++ -o test_exec tests/test.cpp -Iinclude -I/usr/include/eigen3 -std=c++11
-          ./test_exec
+          ./test_exec tests/input_test.json
 
