### 2025-10-10 cf86acc
docs: update README with detailed instructions for running unit tests
