### 2025-10-10 9faf973
update the test.cpp and input.json
