### 2025-10-10 990712a
update run command
