### 2025-10-10 26e6079
update readme.md with clear instructions to compile and run impulsive_start.cpp
