### 2025-10-11 50a9bd5
docs: update README.md with statement of need
