### 2025-10-11 38f641f
docs: Update README.md with statement of need
Signed-off-by: Rohit Chowdhury <d22me102@iitj.ac.in>
