### 2025-10-13 118f6cd
update README.md with a gif
