### 2025-10-14 b07acd6
Implement dynamic file naming in main.cpp for Cl/Cd output
