### 2025-10-15 6585145
docs:(JOSS) update paper - issue #17
