+++ JOSS/paper.bib
@@ -131,6 +131,26 @@ url={https://books.google.co.in/books?id=VnbiEAAAQBAJ}
   title        = {FoilSim III – Interactive Airfoil Simulation},
   author       = {NASA Glenn Research Center},
   year         = {2022},
-  note         = {Available at \url{https://www.grc.nasa.gov/www/k-12/FoilSim3/}}
+  note         = {\url{https://www.grc.nasa.gov/www/k-12/FoilSim3/}}
 }
 
+@article{persson2012numerical,
+  title={Numerical simulation of flapping wings using a panel method and a high-order Navier--Stokes solver},
+  author={Persson, P-O and Willis, David J and Peraire, Jaume},
+  journal={International Journal for Numerical Methods in Engineering},
+  volume={89},
+  number={10},
+  pages={1296--1316},
+  year={2012},
+  publisher={Wiley Online Library},
+  doi={10.1002/nme.3288},
+  url={ https://doi.org/10.1002/nme.3288}
+}
+@inproceedings{chowdhury2025fmfp,
+title = {Development of an unsteady vortex panel method for a flapping airfoil},
+author = {Chowdhury, Rohit and Pathak, Ashish and Arora, Nipun},
+year = {2024},
+month = {06},
+howpublished = {12th International and 52nd National Conference on Fluid Mechanics and Fluid Power (FMFP), 2025, ResearchGate preprint},
+url = {https://www.researchgate.net/publication/393158131_Development_of_an_unsteady_vortex_panel_method_for_a_flapping_airfoil}
+}
