### 2025-10-22 1b2fa39
Update paper.md
Rohit's ORCID id rectified

Signed-off-by: Rohit Chowdhury <d22me102@iitj.ac.in>
