### 2025-10-22 a682ef9
update paper.md with correct ORCID id's
