+++ JOSS/paper.md
@@ -9,7 +9,7 @@ tags:
   - Flapping Wings 
 authors:
   - name: Rohit Chowdhury
-    orcid: 0000-0001-7935-7907
+    orcid: 0009-0003-6701-5477
     affiliation: 1
   - name: Nipun Arora
     orcid: 0000-0002-1835-1189
