+++ JOSS/paper.md
@@ -20,7 +20,7 @@ authors:
 affiliations:
   - index: 1
     name: Department of Mechanical Engineering, Indian Institute of Technology, Jodhpur, Rajasthan, India
-date: 4 April 2025
+date: 22 October 2025
 bibliography: paper.bib
 ---
 
