### 2025-10-22 #8865 @coding4Acause
@piyueh, I have completed the author tasks mentioned, and the latest release of PANKH has been archived on Zenodo. The DOI for this release is: https://doi.org/10.5281/zenodo.17413805. The latest version number of the software, accounting for the changes made during the review process, is  v1.1.0
