### 2025-10-27 e92d0dd
JOSS:paper -fix typos in paper.md and capitalization in paper.bib
