+++ JOSS/paper.bib
@@ -45,7 +45,7 @@ url={https://books.google.co.in/books?id=VnbiEAAAQBAJ}
   author={Katz, Joseph and Plotkin, Allen},
   volume={13},
   year={2001},
-  publisher={Cambridge university press},
+  publisher={{Cambridge university press}},
   doi={10.1115/1.1669432},
   url={https://asmedigitalcollection.asme.org/fluidsengineering/article/126/2/293/458666/LowSpeed-Aerodynamics-Second-Edition}
 }
@@ -102,7 +102,7 @@ url={https://books.google.co.in/books?id=VnbiEAAAQBAJ}
   url={https://doi.org/10.1017/jfm.2017.302}
 }
 @inproceedings{drela1989xfoil,
-  title={XFOIL: An analysis and design system for low Reynolds number airfoils},
+  title={XFOIL: An analysis and design system for low {Reynolds} number airfoils},
   author={Drela, Mark},
   booktitle={Low Reynolds number aerodynamics: proceedings of the conference notre dame, Indiana, USA, 5--7 June 1989},
   pages={1--12},
@@ -135,7 +135,7 @@ url={https://books.google.co.in/books?id=VnbiEAAAQBAJ}
 }
 
 @article{persson2012numerical,
-  title={Numerical simulation of flapping wings using a panel method and a high-order Navier--Stokes solver},
+  title={Numerical simulation of flapping wings using a panel method and a high-order {Navier--Stokes} solver},
   author={Persson, P-O and Willis, David J and Peraire, Jaume},
   journal={International Journal for Numerical Methods in Engineering},
   volume={89},
