### 2025-10-27 87c0505
update paper.bib by fixing capitalization of certain words
