### 2025-10-27 9f08f6e
update paper.bib by fixing capitalization of certain words
