+++ JOSS/paper.bib
@@ -45,7 +45,7 @@ url={https://books.google.co.in/books?id=VnbiEAAAQBAJ}
   author={Katz, Joseph and Plotkin, Allen},
   volume={13},
   year={2001},
-  publisher={{Cambridge} {university} {press}},
+  publisher={{Cambridge University Press}},
   doi={10.1115/1.1669432},
   url={https://asmedigitalcollection.asme.org/fluidsengineering/article/126/2/293/458666/LowSpeed-Aerodynamics-Second-Edition}
 }
