### 2025-10-28 #8865 @piyueh
> I think it's this line from the paper:
>
> ```
> The resulting system, formulated as shown in **Figure 2**, is expressed as a single matrix equation and solved using the
>  PartialPivLU decomposition solver available in the *Dense Linear Problems and Decompositions* module of the Eigen
> Library [[@eigenweb]](https://eigen.tuxfamily.org/index.php?title=Main_Page). In-depth insights into post-processing
> steps, encompassing the solution of the unsteady Bernoulli equation and computations of aerodynamic loads, are
> detailed in [@chowdhury2025fmfp].
> ```
>
> In particular: `[[@eigenweb]](https://eigen.tuxfamily.org/index.php?title=Main_Page)`
>
> Is `@eigenweb` a citation (e.g., is it in the references) or a link?

@coding4Acause See the quoted message 👆 . Please change `[[@eigenweb]](https://eigen.tuxfamily.org/index.php?title=Main_Page)` to just a citation. The key `eigenweb` is already in `paper.bib`.
