### 2025-10-28 44df259
docs:(JOSS) update paper.md with correct citation for eigen lib.
