### 2025-10-28 #8865 @piyueh
@coding4Acause I've recommended acceptance, and the editor in chief will take over from this point. In the mean time, you can check the final proof as described [here](https://github.com/openjournals/joss-reviews/issues/8865#issuecomment-3457182973).

@andrewgiuliani, @stebiondi, @daavid00 Thanks for reviewing this submission for JOSS. I really appreciate your time!
