### 2025-11-04 98c26fe
JOSS paper edits
Signed-off-by: Kyle Niemeyer <kyleniemeyer@fastmail.com>
