### 2025-11-04 #8865 @kyleniemeyer
@coding4Acause I've made some edits in https://github.com/coding4Acause/PANKH/pull/20, please review and merge.

In addition, please update the Chowdhury et al. citation (it currently points to a ResearchGate URL).
