### 2025-11-05 d3bcbee
update paper.bib -fix citation and paper.md
