+++ JOSS/paper.bib
@@ -101,16 +101,6 @@ url={https://books.google.co.in/books?id=VnbiEAAAQBAJ}
   doi={10.1017/jfm.2017.302},
   url={https://doi.org/10.1017/jfm.2017.302}
 }
-@inproceedings{drela1989xfoil,
-  title={XFOIL: An analysis and design system for low {Reynolds} number airfoils},
-  author={Drela, Mark},
-  booktitle={Low Reynolds number aerodynamics: proceedings of the conference notre dame, Indiana, USA, 5--7 June 1989},
-  pages={1--12},
-  year={1989},
-  organization={Springer},
-  doi={10.1007/978-3-642-84010-4_1},
-  url={https://link.springer.com/chapter/10.1007/978-3-642-84010-4_1}
-}
 
 @misc{eigenweb,
   author       = {Gaël Guennebaud and Benoît Jacob and others},
@@ -146,11 +136,23 @@ url={https://books.google.co.in/books?id=VnbiEAAAQBAJ}
   doi={10.1002/nme.3288},
   url={ https://doi.org/10.1002/nme.3288}
 }
+
+@inproceedings{drela1989xfoil,
+  title={XFOIL: An analysis and design system for low {Reynolds} number airfoils},
+  author={Drela, Mark},
+  booktitle={Low Reynolds number aerodynamics: proceedings of the conference notre dame, Indiana, USA, 5--7 June 1989},
+  pages={1--12},
+  year={1989},
+  organization={Springer},
+  doi={10.1007/978-3-642-84010-4_1},
+  url={https://link.springer.com/chapter/10.1007/978-3-642-84010-4_1}
+}
+
 @inproceedings{chowdhury2025fmfp,
-title = {Development of an unsteady vortex panel method for a flapping airfoil},
-author = {Chowdhury, Rohit and Pathak, Ashish and Arora, Nipun},
-year = {2024},
-month = {06},
-howpublished = {12th International and 52nd National Conference on Fluid Mechanics and Fluid Power (FMFP), 2025, ResearchGate preprint},
-url = {https://www.researchgate.net/publication/393158131_Development_of_an_unsteady_vortex_panel_method_for_a_flapping_airfoil}
+  author    = {Chowdhury, Rohit and Pathak, Ashish and Arora, Nipun},
+  title     = {Development of an Unsteady Vortex Panel Method for a Flapping Airfoil},
+  booktitle = {Proceedings of the 12th International and 52nd National Conference on Fluid Mechanics and Fluid Power (FMFP 2025)},
+  year      = {2025},
+  address   = {Aligarh,India},  % Replace with actual location if known
+  note      = {Accepted for publication},
 }
