### 2025-11-05 ec50b57
docs: fix PANKH acronym expansion in README (of → for)
