### 2025-11-10 87de6d9
Update paper.bib
Fix the Bib entry for Chowdhury2025fmfp.

Signed-off-by: Rohit Chowdhury <d22me102@iitj.ac.in>
