+++ JOSS/paper.bib
@@ -153,6 +153,7 @@ url={https://books.google.co.in/books?id=VnbiEAAAQBAJ}
   title     = {Development of an Unsteady Vortex Panel Method for a Flapping Airfoil},
   booktitle = {Proceedings of the 12th International and 52nd National Conference on Fluid Mechanics and Fluid Power (FMFP 2025)},
   year      = {2025},
-  address   = {Aligarh,India},  % Replace with actual location if known
+  address   = {Aligarh,India},
   note      = {Accepted for publication},
+
 }
