### 2025-11-10 be5125c
Update paper.bib
Signed-off-by: Rohit Chowdhury <d22me102@iitj.ac.in>
