+++ JOSS/paper.bib
@@ -154,6 +154,6 @@ url={https://books.google.co.in/books?id=VnbiEAAAQBAJ}
   booktitle = {Proceedings of the 12th International and 52nd National Conference on Fluid Mechanics and Fluid Power (FMFP 2025)},
   year      = {2025},
   address   = {Aligarh,India},
-  note      = {Accepted for publication},
-
+  note      = {Accepted for publication}
 }
+
